requirements.txt
click==6.7
flake8==3.5.0
Flask==1.0.2
Flask-Testing==0.7.1
Flask-WTF==0.14.2
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
mccabe==0.6.1
pycodestyle==2.3.1
pyflakes==1.6.0
python-dotenv==0.8.2
Werkzeug==0.14.1
WTForms==2.2.1
